+++ crates/admin/src/cluster_controller/mod.rs
@@ -9,6 +9,7 @@
 // by the Apache License, Version 2.0.
 
 pub mod cluster_state_refresher;
+pub mod grpc_svc_handler;
 mod logs_controller;
 mod observed_cluster_state;
 pub mod protobuf;
