+++ crates/node/src/network_server/grpc_svc_handler.rs
@@ -8,57 +8,51 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use arrow_flight::encode::FlightDataEncoderBuilder;
+// use arrow_flight::encode::FlightDataEncoderBuilder;
 use arrow_flight::error::FlightError;
 use enumset::EnumSet;
 use futures::stream::BoxStream;
-use futures::TryStreamExt;
 use tokio_stream::StreamExt;
 use tonic::{Request, Response, Status, Streaming};
 
 use restate_core::network::protobuf::node_svc::node_svc_server::NodeSvc;
 use restate_core::network::protobuf::node_svc::IdentResponse;
 use restate_core::network::protobuf::node_svc::{StorageQueryRequest, StorageQueryResponse};
-use restate_core::network::ProtocolError;
-use restate_core::network::{ConnectionManager, GrpcConnector};
+use restate_core::network::ConnectionManager;
+use restate_core::network::{ProtocolError, TransportConnect};
 use restate_core::{metadata, TaskCenter};
 use restate_types::health::Health;
 use restate_types::nodes_config::Role;
 use restate_types::protobuf::node::Message;
 
-use crate::network_server::WorkerDependencies;
-
-pub struct NodeSvcHandler {
+pub struct NodeSvcHandler<T> {
     task_center: TaskCenter,
     cluster_name: String,
     roles: EnumSet<Role>,
     health: Health,
-    worker: Option<WorkerDependencies>,
-    connections: ConnectionManager<GrpcConnector>,
+    connections: ConnectionManager<T>,
 }
 
-impl NodeSvcHandler {
+impl<T: TransportConnect> NodeSvcHandler<T> {
     pub fn new(
         task_center: TaskCenter,
         cluster_name: String,
         roles: EnumSet<Role>,
         health: Health,
-        worker: Option<WorkerDependencies>,
-        connections: ConnectionManager<GrpcConnector>,
+        connections: ConnectionManager<T>,
     ) -> Self {
         Self {
             task_center,
             cluster_name,
             roles,
             health,
-            worker,
             connections,
         }
     }
 }
 
 #[async_trait::async_trait]
-impl NodeSvc for NodeSvcHandler {
+impl<T: TransportConnect> NodeSvc for NodeSvcHandler<T> {
     async fn get_ident(&self, _request: Request<()>) -> Result<Response<IdentResponse>, Status> {
         let node_status = self.health.current_node_status();
         let admin_status = self.health.current_admin_status();
@@ -90,36 +84,37 @@ impl NodeSvc for NodeSvcHandler {
 
     async fn query_storage(
         &self,
-        request: Request<StorageQueryRequest>,
+        _request: Request<StorageQueryRequest>,
     ) -> Result<Response<Self::QueryStorageStream>, Status> {
-        let Some(ref worker) = self.worker else {
-            return Err(Status::failed_precondition("Not a worker node"));
-        };
-        let query = request.into_inner().query;
-
-        let record_stream = self
-            .task_center
-            .run_in_scope("query-storage", None, async move {
-                worker.query_context.execute(&query).await.map_err(|err| {
-                    Status::internal(format!("failed executing the query '{}': {}", query, err))
-                })
-            })
-            .await?;
-
-        let schema = record_stream.schema();
-        let response_stream =
-            FlightDataEncoderBuilder::new()
-                // CLI is expecting schema information
-                .with_schema(schema)
-                .build(record_stream.map_err(|err| {
-                    FlightError::from(datafusion::arrow::error::ArrowError::from(err))
-                }))
-                .map_ok(|flight_data| StorageQueryResponse {
-                    header: flight_data.data_header,
-                    data: flight_data.data_body,
-                })
-                .map_err(flight_error_to_tonic_status);
-        Ok(Response::new(Box::pin(response_stream)))
+        // let Some(ref worker) = self.worker else {
+        //     return Err(Status::failed_precondition("Not a worker node"));
+        // };
+        // let query = request.into_inner().query;
+        //
+        // let record_stream = self
+        //     .task_center
+        //     .run_in_scope("query-storage", None, async move {
+        //         worker.query_context.execute(&query).await.map_err(|err| {
+        //             Status::internal(format!("failed executing the query '{}': {}", query, err))
+        //         })
+        //     })
+        //     .await?;
+        //
+        // let schema = record_stream.schema();
+        // let response_stream =
+        //     FlightDataEncoderBuilder::new()
+        //         // CLI is expecting schema information
+        //         .with_schema(schema)
+        //         .build(record_stream.map_err(|err| {
+        //             FlightError::from(datafusion::arrow::error::ArrowError::from(err))
+        //         }))
+        //         .map_ok(|flight_data| StorageQueryResponse {
+        //             header: flight_data.data_header,
+        //             data: flight_data.data_body,
+        //         })
+        //         .map_err(flight_error_to_tonic_status);
+        // Ok(Response::new(Box::pin(response_stream)))
+        todo!()
     }
 
     type CreateConnectionStream = BoxStream<'static, Result<Message, Status>>;
